% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{active_initial_messages}
\alias{active_initial_messages}
\title{Prints initial messages for active sampling, if needed.}
\usage{
active_initial_messages(n_cluster, query_type)
}
\description{
Prints initial messages for active sampling, if needed.
}
