% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active.R
\name{active_label}
\alias{active_label}
\title{Active Learning EM Algorithm}
\usage{
active_label(
  docs,
  labels = c(0, 1),
  lambda = 1,
  n_class = 2,
  n_cluster = 2,
  max_active = 5,
  init_size = 10,
  max_query = 10,
  save_file_name = NA,
  save_directory = NA
)
}
\arguments{
\item{docs}{[matrix]     Matrix of labeled and unlabeled documents, where each
row has index values and a nested Matrix of word tokens.}

\item{labels}{[vector]     Vector of character strings indicating classification
options for labeling.}

\item{lambda}{[numeric]    Numeric value between 0 and 1. Used to weight unlabeled
documents.}

\item{n_class}{[numeric]    Number of classes to be considered.}

\item{max_active}{[numeric]    Value of maximum allowed active learning iterations.}

\item{init_size}{[numeric]    Value of maximum allowed iterations within the EM algorithm.}

\item{max_query}{[numeric]    Maximum number of documents queried in each EM iteration.}

\item{doc_name}{[character]  Character string indicating the variable in 'docs'
that denotes the text of the documents to be classified.}
}
\value{
[list]       List containing labeled document matrix, prior weights, word likelihoods, and a vector
                                 of user-labeled documents ids.
}
\description{
Active learning for weighted-EM algorithm. After initial EM algorithm converges,
              oracle is queried for labels to documents that the EM algorithm was most unsure of.
              This process iterates until max iterations are reached, or there are no
              documents in the window of uncertainty.
}
