% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active.R
\name{active_label_from_saved}
\alias{active_label_from_saved}
\title{Active Learning EM Algorithm}
\usage{
active_label_from_saved(saved_file)
}
\arguments{
\item{saved_file}{[string]     Name of saved file}
}
\description{
Active learning for weighted-EM algorithm. After initial EM algorithm converges,
              oracle is queried for labels to documents that the EM algorithm was most unsure of.
              This process iterates until max iterations are reached, or there are no
              documents in the window of uncertainty.
}
