% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{agg_helper_convert}
\alias{agg_helper_convert}
\title{Aggregation Helper}
\usage{
agg_helper_convert(model_preds, n_cluster_collapse_type = "simple")
}
\description{
helps aggregation function by collapsing clusters to classes (binary only).
}
