% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{aggregate_model_predictions}
\alias{aggregate_model_predictions}
\title{Aggregate Model Predictions}
\usage{
aggregate_model_predictions(
  pred_lst,
  dfms = NULL,
  models = NULL,
  val_data = NULL,
  n_cluster,
  agg_type = "random",
  n_cluster_collapse_type = "simple"
)
}
\description{
Processes model predictions according to cluster structure,
then chooses between model predictions for each dfm being
used to fit a model.
}
