% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{balancing_data}
\alias{balancing_data}
\title{Create a (un)balanced dataset with specified size and ratio of positive documents}
\usage{
balancing_data(data, pos_ratio, size = nrow(data), seed = 1990)
}
\arguments{
\item{pos_ratio}{ratio of positive document ()}

\item{iter}{the number of sampling}
}
\description{
Create a (un)balanced dataset with specified size and ratio of positive documents
}
