% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{check_extreme_words_inlabel}
\alias{check_extreme_words_inlabel}
\title{Check how often the extreme words are in the labeled documents.}
\usage{
check_extreme_words_inlabel(
  data,
  docsType,
  docName,
  index_name,
  labelsName,
  dfm,
  n_extreme = 10,
  iter = 1000,
  sample_size = 100,
  sampling_method = NULL
)
}
\arguments{
\item{n_extreme}{number of extreme words}

\item{iter}{the number of sampling}
}
\description{
check the distribution of the extreme words in the labeled documents of size 100
}
