% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{check_lr_convergence}
\alias{check_lr_convergence}
\title{check for convergence if using log-ratio sampling}
\usage{
check_lr_convergence(output, count, log_ratio_threshold, log_ratio_conv_type)
}
\description{
check for convergence if using log-ratio sampling
}
