% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{choose_best_model}
\alias{choose_best_model}
\title{Choose best model.}
\usage{
choose_best_model(model_preds, model_weights, index_name = "id")
}
\description{
gets predictions from single best model
}
