% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{classification_gui}
\alias{classification_gui}
\title{GUI to select labels.}
\usage{
classification_gui(options, documents, param_to_save, selections = NA)
}
\arguments{
\item{options}{[vector]    Vector of string labels}

\item{documents}{[vector]    Vector of string texts.}

\item{param_to_save}{[list]    List of variables to save if user chooses to.}

\item{selections}{[vector]   Already labeled documents}
}
\description{
Helps select labels.
}
