% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{clean_data}
\alias{clean_data}
\title{Structure Data}
\usage{
clean_data(
  docs,
  n_class,
  doc_name,
  index_name,
  labels_name = NULL,
  filters = NULL,
  add_index = T,
  add_filter = T,
  keep_labels = F
)
}
\arguments{
\item{docs}{[matrix]      Matrix of labeled and/or unlabeled documents.}

\item{n_class}{[numeric]     Number of classes to be considered.}

\item{doc_name}{[string]      Character string indicating the variable in 'docs'
that denotes the text of the documents to be classified.}

\item{index_name}{[character]   Character string indicating the variable in 'docs' that
denotes the index value of the document to be classified.}

\item{labels_name}{[character]   Character string indicating the variable in \code{docs}
that denotes the already known labels of the documents.
By default, value is set to \code{NULL}.}

\item{filters}{[character]   A vector of regular expressions used to filter out unwanted documents.}

\item{add_index}{[logical]     Boolean logical value indicating whether or not add an index
in the restructuring process.}

\item{add_filter}{[logical]     Boolean logical value indicating whether or not to filter documents
in the restructuring process.}

\item{keep_labels}{[logical]     Boolean logical value indicating whether or not to keep an existing column
of labels in the dataset.}
}
\value{
[matrix]      Structured matrix of labeled and unlabeled documents, updated with
                                labels for the documents in 'toLabel'.
}
\description{
Structures data to prepare for Active-EM implementation.
                 Options to filter documents by chosen character strings, as well as to
                 add index value for each document.
}
