% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{compare_model_results}
\alias{compare_model_results}
\title{Compare Model Results}
\usage{
compare_model_results(docs, docsType, querySize, fixed_words = NULL, ...)
}
\description{
Run models according to chosen specifications for fakenews data.
}
