% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EM_helper.R
\name{E_step}
\alias{E_step}
\title{E step of the EM algorithm}
\usage{
E_step(
  .D_test,
  .X_c = NA,
  .X_b = NA,
  .class_prob,
  .word_prob,
  .mu = NA,
  .sig = NA,
  .psi = NA
)
}
\arguments{
\item{.D_test}{document term matrix for the test documents}

\item{.class_prob}{log of the class probability}

\item{.word_prob}{log of the word probability}
}
\value{
a matrix of the log probability of each doc having each class 
dimension is length(D_test) by 2
}
\description{
E step of the EM algorithm
}
