% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EM_helper.R
\name{E_step_multi}
\alias{E_step_multi}
\title{E step with multiple cluster}
\usage{
E_step_multi(
  .C_train,
  .D_train,
  .D_test,
  .X_c = NA,
  .X_b = NA,
  .class_prob,
  .word_prob,
  .mu = NA,
  .psi = NA,
  .sig = NA
)
}
\arguments{
\item{.D_test}{document term matrix for the test data}

\item{.class_prob}{log of the class probability}

\item{.word_prob}{log of the word probability}

\item{.mu}{log of continuous metadata probability}

\item{.psi}{log of binary metadata probability}

\item{.C_train:}{class matrix of the training data}

\item{.D_train:}{document term matrix of the training data}
}
\value{
a matrix of the log probability of each doc having each class 
dimension is length(D_test) by 2
Vectorize matrices, stack them, apply logsumexp,
then reshape into matrix form.
}
\description{
E step of the EM algorithm with multiple clusters
it has to take .D_train (labeled documents) too because
for negative label documents, we have to estimate cluster probability
TO BE UPDATED
}
