% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{gen_results_tbl}
\alias{gen_results_tbl}
\title{Generate Results Table}
\usage{
gen_results_tbl(include_out_stats, metadata, max_iters, model_name)
}
\description{
generates an object for storing model results
}
