% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{generate_lambda_vec}
\alias{generate_lambda_vec}
\title{Creates a vector of lambdas depending on decay value}
\usage{
generate_lambda_vec(lambda_decay, lambda, rate, iters, max_active)
}
\description{
Creates a vector of lambdas depending on decay value
}
