% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{get_class_matrix}
\alias{get_class_matrix}
\title{Get Document Class Probability Matrix}
\usage{
get_class_matrix(
  docs,
  n_class,
  hand_labeled_index,
  doc_name,
  index_name,
  NB_init
)
}
\arguments{
\item{docs}{[matrix]    Matrix of labeled and unlabeled documents.}

\item{n_class}{[numeric]   Number of classes to be considered.}

\item{hand_labeled_index}{[vector]    Vector of index values for hand labeled documents in \code{docs}.}

\item{doc_name}{[character] Character string indicating the variable in 'docs'
that denotes the text of the documents .}

\item{index_name}{[character] Character string indicating the variable in 'docs'
that denotes the index value of the documents .}
}
\value{
[matrix]    Class assignment matrix.
}
\description{
Gets matrix of document classifications.
}
