% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EM_helper.R
\name{get_class_prob}
\alias{get_class_prob}
\title{get weighted log of class probability}
\usage{
get_class_prob(.D_train, .D_test, .E, .n_class, .lambda, .supervise)
}
\arguments{
\item{.D_train}{document term matrix for the training documents}

\item{.D_test}{document term matrix for the test documents}

\item{.E}{output of E step for all documents}

\item{.n_class}{number of classes}

\item{.lambda}{weight}

\item{supervise}{T if supervise, F is unsupervise}
}
\value{
a matrix of log likelihood of a class(nrow = 2, ncol = # of document)
}
\description{
get weighted log of class probability
}
