% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EM_helper.R
\name{get_class_prob_NB}
\alias{get_class_prob_NB}
\title{NB step, class probability}
\usage{
get_class_prob_NB(.D_train, .C_train)
}
\arguments{
\item{.D_train}{DTM for training docs}

\item{.C_train}{class matrix for training docs}
}
\description{
NB step, class probability
}
