% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{get_classes}
\alias{get_classes}
\title{Get Classes}
\usage{
get_classes(n_class)
}
\arguments{
\item{n_class}{[numeric] Number of classes to be considered.}
}
\value{
[vector] Vector of classes.
}
\description{
Builds a vector of classification options.
}
