% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{get_clusters}
\alias{get_clusters}
\title{Get Clusters}
\usage{
get_clusters(n_cluster)
}
\arguments{
\item{n_class}{[numeric] Number of clusters to be considered.}
}
\value{
[vector] Vector of clusters.
}
\description{
Builds a vector of classification options.
}
