% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{get_conf_matrix_weight}
\alias{get_conf_matrix_weight}
\title{Obtain weighted False negative and false positive in the Confusio nmatrix}
\usage{
get_conf_matrix_weight(
  docs,
  labelsName,
  index_name = NULL,
  labeledIndex = NULL,
  levels = c(0, 1),
  n_class = 2,
  n_cluster = 2,
  useLabeled = T
)
}
\description{
Obtain weighted False negative and false positive in the Confusio nmatrix
}
