% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{get_dfm}
\alias{get_dfm}
\title{Get Document Feature Matrix}
\usage{
get_dfm(
  docs,
  doc_name = "text",
  index_name = "id",
  stem = T,
  ngrams = 1,
  trimPct = 1e-04,
  min_doc_freq = 2,
  idfWeight = F,
  removeStopWords = T,
  minChar = 4
)
}
\arguments{
\item{docs}{[matrix]    Matrix of labeled and unlabeled documents.}

\item{doc_name}{[character] Character string indicating the variable in 'docs'
that denotes the text of the documents to be classified.}

\item{index_name}{[character] Character string indicating the variable in 'docs'
that denotes the index value of the document to be classified.}

\item{stem}{[logical]   Switch indicating whether or not to stem terms.}

\item{ngrams}{[integer]   Integer value indicating the size of the ngram to use to build the dfm.}

\item{trimPct}{[numeric]   Numeric value indicating the threshold of percentage of document
membership at which to remove terms from the data-term matrix.
E.g., if \code{trimPct = .5}, then all words that are in less than
50 percent of the documents will be removed.}

\item{min_doc_freq}{[integer]   Minimum number of documents a term must be in to stay in the document term matrix.}

\item{idfWeight}{[logical]   Switch indicating whether to weight the document term matrix by the frequency of
word counts. Only works if \code{dfmType = "quanteda"}.}
}
\value{
[matrix]    Document term matrix.
}
\description{
Builds document feature matrix using quanteda package.
}
