% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{get_empirical_dist}
\alias{get_empirical_dist}
\title{Get True Empirical Word Distributions.}
\usage{
get_empirical_dist(
  docs,
  docsType,
  doc_name,
  index_name,
  labelsName,
  n_extreme = 50,
  dfm = NULL,
  extreme_type = "abs_diff",
  ...
)
}
\description{
Get data showing the true empirical distribution of word likelihoods from labeled data.
}
