% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{get_entropy}
\alias{get_entropy}
\title{Get Entropy Value}
\usage{
get_entropy(data)
}
\arguments{
\item{[matrix]}{Matrix of data.}
}
\value{
[vector] Vector of entropy values.
}
\description{
Gets row-wise entropy values from a rectangular data array.
}
