% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{get_figures_grid}
\alias{get_figures_grid}
\title{Get Figures}
\usage{
get_figures_grid(
  output,
  ncol = 2,
  which_measures = "all",
  which_models = "all",
  labels = F,
  fixed_y_axis = F,
  dataset_name = NA
)
}
\arguments{
\item{output}{List; output of `run_models_fast`.}

\item{ncol}{Integer; number of columns for figure.}

\item{which_measures}{Vector of string values indicating which measures to show in the figure.
By default, set to "all", but can take a vector like `c("F1_out", "acc_out")` to only see the results
for those two measures.}

\item{which_models}{Vector of string values indicating which models to show in the figure.
By default, set to "all", but can take a vector like `c("aNB", "aEM", "awEM")`, for example,
to only see from those three models.}

\item{labels}{Boolean; indicates whether to access metadata from `output` for figure labels.}

\item{fixed_y_axis}{Boolean; indicates whether to fix y-axis scale from 0 to 1.}

\item{dataset_name}{String; Used to generate title for figure.}
}
\value{
Plotted values of chosen measures across experiments.
}
\description{
Turns output from `run_models_fast` into graphical output for observation of results.
}
