% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{get_fixed_words_mtx}
\alias{get_fixed_words_mtx}
\title{Get a Matrix of Fixed Words and Class Probabilities.}
\usage{
get_fixed_words_mtx(data)
}
\description{
Gets a matrix of extreme words and respective word probabilities. Used to
supply to EM algorithm to fix word probabilities.
}
