% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{get_index}
\alias{get_index}
\title{Get Index}
\usage{
get_index(docs, index_name)
}
\arguments{
\item{docs}{[numeric]   Documents chosen to initialize system.}

\item{index_name}{[character] Character string indicating the variable in 'docs'
that denotes the index value of the document to be classified.}
}
\value{
[vector] Vector of classes.
}
\description{
Gets index from a subset of doucments.
}
