% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{get_initial_documents}
\alias{get_initial_documents}
\title{gets initial documents for active learning algorithm}
\usage{
get_initial_documents(docs, init_index, index_name, init_size, whichOutTest)
}
\description{
gets initial documents for active learning algorithm
}
