% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{get_label_prop_fig}
\alias{get_label_prop_fig}
\title{Get Label Proportion Figure}
\usage{
get_label_prop_fig(out, dataset_name, index_name = "id")
}
\description{
Produces a figure that proportion of positive to negative
queries across active iteration.
}
