% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EM_helper.R
\name{get_maximand}
\alias{get_maximand}
\title{get maximand. Use this to check if the EM keeps incrasing maximand}
\usage{
get_maximand(
  .D_train,
  .D_test,
  .E,
  .X_c_train,
  .X_c_test,
  .X_b_train,
  .X_b_test,
  .class_prob,
  .word_prob,
  .mu = NA,
  .psi = NA,
  .sig = NA,
  supervise
)
}
\arguments{
\item{.D_train}{document term matrix for the training documents}

\item{.D_test}{document term matrix for the test documents}

\item{.E}{output of E step for all documents}

\item{.class_prob}{output of get_class_prob()}

\item{.word_prob}{output of get_word_prob()}

\item{.supervise}{T if supervised. F is unsupervised}
}
\value{
a scalar of the value of maximand in each EM iteration
}
\description{
get maximand. Use this to check if the EM keeps incrasing maximand
}
