% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{get_mean_mpe}
\alias{get_mean_mpe}
\title{Get Mean Prediction Error Singular}
\usage{
get_mean_mpe(
  mod,
  dfm,
  val_data,
  labels_name = "label",
  index_name = "id",
  n_cluster_collapse_type = "simple",
  n_cluster
)
}
\description{
gets mean prediction error
}
