% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{get_mean_mpes}
\alias{get_mean_mpes}
\title{Get Mean Prediction Error}
\usage{
get_mean_mpes(dfms, models, val_data, n_cluster)
}
\description{
gets mean mpes across a list of dfms and list of models of equal length
}
