% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{get_model_weights}
\alias{get_model_weights}
\title{Get Model Weights}
\usage{
get_model_weights(dfms, models, val_data, n_cluster)
}
\description{
calculates the weights that each model recieves and normalize
}
