% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{get_predict_prop_fig}
\alias{get_predict_prop_fig}
\title{Get Prediction Probability Figure}
\usage{
get_predict_prop_fig(
  out,
  dataset_name,
  index_name = "id",
  doc_labels = FALSE,
  metadata = TRUE,
  arrange_y_axis = FALSE,
  out_sample = TRUE,
  labeled_only = FALSE
)
}
\description{
Produces a figure that charts changes to out of sample
classification probability across active iterations.
}
