% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{get_results_matrix}
\alias{get_results_matrix}
\title{Get Results Matrix for active_EM Output.}
\usage{
get_results_matrix(out, out_sample = FALSE, model_name)
}
\arguments{
\item{out}{list; Output from `active_EM` function.}

\item{out_sample}{logical; Indicates whether to get output matrix
out of sample test. Will only work if 'active_EM' exports an `out_docs`
object.}

\item{model_name}{string; Name of model.}
}
\value{
mtx; Object for feeding into `get_figure_grid` function.
}
\description{
Translates activeEM() output object
into result matrix for graphing.
}
\note{
Ensure that 'exportAll = T' in `active_EM` function.
}
