% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{get_uncertain_docs}
\alias{get_uncertain_docs}
\title{Get Uncertain Documents}
\usage{
get_uncertain_docs(
  docs,
  bound,
  max_query,
  index_name,
  hand_labeled_index,
  force_list = F,
  query_type = "basic_entropy",
  quantileBreaks = c(75, 20),
  sampleProps = c(0.5, 0.3, 0.2),
  mu = 0.001,
  tau = 0.001,
  regions = "both",
  dfm = NULL,
  seed = NULL,
  n_cluster = NULL
)
}
\arguments{
\item{docs}{[matrix]    Matrix of labeled and unlabeled documents.}

\item{bound}{[numeric]   The choice of lower bound for entropy-based uncertainty selection.}

\item{max_query}{[numeric]   Maxmium number of uncertain documents that can be queried.}

\item{index_name}{[character] Character string indicating the variable in 'docs'
that denotes the index value of the documents .}

\item{hand_labeled_index}{[vector]    Vector of index values for hand labeled documents in \code{docs}.}

\item{force_list}{[logical]   Switch indicating whether to force the filtering of documents with
no entropy. Set to \code{FALSE} by default.}

\item{query_type}{[string]    String indicating which type of uncertainty sampling to use. Options are \code{"standard_entropy"},
\code{"normalized_entropy"}, \code{"tiered_entropy"}, or
\code{"tiered_entropy_weighted"}.}

\item{quantileBreaks}{[vector]    Vector of break points to distinguish entropy zones. The first value is
the break point between the first and second tier, the second is the
break point between the second and third tier.}

\item{sampleProps}{[vector]    Vector of sampling proportions for each entropy zone. The first value is
the proportion of \code{max_query} to be sampled from the high entropy region,
the second value is the proportion to be sampled from the middle entropy region,
and the third value is the proportion to be sampled from the lowest entropy region.}

\item{n_cluster}{[int]       Number of clusters.}
}
\value{
[vector]    Vector of id values of documents that the EM algorithm is uncertain about.
}
\description{
Get documents that the previous iteration of the EM algorithm is least sure about.
}
