% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{get_weighted_prediction}
\alias{get_weighted_prediction}
\title{Get Weighted Prediction}
\usage{
get_weighted_prediction(
  model_preds,
  model_weights,
  index_name = "id",
  labels_name = "label"
)
}
\description{
gets weighted predictions for unlabled documents
}
