% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EM_helper.R
\name{get_word_prob}
\alias{get_word_prob}
\title{get log likelihood of a document}
\usage{
get_word_prob(.D, .E, .beta = NA, .fixed_words = NULL)
}
\arguments{
\item{.D}{document term matrix}

\item{.E}{labels for all documents}

\item{.beta}{prior parameter for eta}

\item{.fixed_words}{matrix of fixed words with class probabilities,
where ncol is the number of classes.}
}
\value{
a matrix of word probability(nrow = 2, ncol = V (size of vocab))
}
\description{
get log likelihood of a document
}
