% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EM_helper.R
\name{get_word_prob_log}
\alias{get_word_prob_log}
\title{get log likelihood of a document}
\usage{
get_word_prob_log(.C_train, .D_train, .D_test, .E, .n_class, .lambda)
}
\arguments{
\item{.E}{output from the E step. Log scale.}

\item{.D}{document term matrix}

\item{.fixed_words}{matrix of fixed words with class probabilities,
where ncol is the number of classes.}
}
\value{
a matrix of word probability(nrow = 2, ncol = V (size of vocab))
}
\description{
get log likelihood of a document
}
