% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EM_helper.R
\name{get_word_prob_NB}
\alias{get_word_prob_NB}
\title{Get Word Probability (Naive Bayes)}
\usage{
get_word_prob_NB(.D_train, .C_train, .beta = NA)
}
\arguments{
\item{.D_train}{DTM for training docs}

\item{.C_train}{class matrix for training docs}

\item{.beta}{prior parameter for eta (positive column)}
}
\description{
get word probability for NB step
}
