% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{initialize_beta_tbl}
\alias{initialize_beta_tbl}
\title{initializes a table of prior values for eta}
\usage{
initialize_beta_tbl(dfms, n_class, keywords_list = NA, gamma)
}
\description{
initializes a table of prior values for eta
}
