% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{KL}
\alias{KL}
\title{KL Divergence Function}
\usage{
KL(p, q)
}
\arguments{
\item{p}{Empirical distribution.}

\item{q}{Posterior distribution.}
}
\description{
Gets KL divergence between two distributions.
}
