% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{log_ratio_sample}
\alias{log_ratio_sample}
\title{Log-Ratio Sampling}
\usage{
log_ratio_sample(docs, out, dfm, mu, tau, max_query, edge, regions)
}
\arguments{
\item{docs}{[dataframe] Documents with log ratio and cumulative sum columns.}

\item{out}{[list] List of output objects from `EM()` function.}

\item{dfm}{[quanteda dfm] Quanteda dfm object.}

\item{mu}{Error acceptance rate for first region.}

\item{max_query}{[integer] Number of documents to be queried.}

\item{edge}{[logical] Whether we sample from documents that have very high/low posterior}

\item{eta}{Error acceptance rate for second region.}
}
\value{
Sampled documents.
}
\description{
Log-Ratio Sampling
}
