% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{match_clusters_to_docs}
\alias{match_clusters_to_docs}
\title{Match Multicluster EM Output to Document Matrix}
\usage{
match_clusters_to_docs(docs, EMoutput, index_name, n_cluster)
}
\value{
Matrix of documents.
}
\description{
Matches the output of multicluster EM to the document corpus matrix.
}
