% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{match_EM_to_docs}
\alias{match_EM_to_docs}
\title{Match EM Output to Document Matrix}
\usage{
match_EM_to_docs(
  docs,
  EMoutput,
  classes,
  doc_name,
  index_name,
  labels_name = NULL
)
}
\arguments{
\item{docs}{[matrix]    Matrix of labeled and unlabeled documents.}

\item{EMoutput}{[matrix]    Matrix of expected class assignments produced by EM algorithm..}

\item{classes}{[vector]    Vector of character strings indicating the classes being considered.}

\item{doc_name}{[character]}

\item{index_name}{[character] Character string indicating the variable in 'docs'
that denotes the index value of the documents .}

\item{labels_name}{[character] Character string indicating the variable in \code{docs}
that denotes the already known labels of the documents.
By default, value is set to \code{NULL}.}
}
\value{
[matrix]    Matrix of documents that the EM algorithm is uncertain about.
}
\description{
Matches the output of the previous run of the EM algorithm to the matrix of documents.
}
