% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{multi_fig_analysis}
\alias{multi_fig_analysis}
\title{Multiple Figure Analyis of Unsupervised EM}
\usage{
multi_fig_analysis(
  docs,
  docsType,
  initType,
  docName,
  index_name,
  labelsName,
  extreme_type = "abs_diff",
  stem = T,
  trimPct = 1e-04,
  minDocFreq = 2,
  idfWeight = F,
  removeStopWords = T,
  minChar = 4,
  ngrams = 1
)
}
\description{
Runs unsupervised EM and summarises output in three figures.
}
