% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{query_eta}
\alias{query_eta}
\title{Query Eta}
\usage{
query_eta(
  eta,
  class,
  n_query,
  handlabel = TRUE,
  true_eta = NA,
  keyword_select_scheme = "ratio",
  val_scheme = "ratio",
  verbose = FALSE,
  dfm,
  keywords_list = NA
)
}
\arguments{
\item{eta}{Eta parameter from EM.}

\item{class}{Class under consideration.}

\item{n_query}{Number of words to query.}

\item{handlabel}{Boolean indicating whether or not to label by hand.}

\item{true_eta}{If handlabel is false and an eta matrix is provided
to the true_eta param, values from the true_eta are used to determine
whether or not a word should be flagged.}

\item{keyword_select_scheme}{Keyword selection scheme. Either "ranked"
for top eta keyword selection, "ratio" for top eta ratio keyword selection,
or "combined" for log word count + log eta ratio selection.}

\item{val_scheme}{If handlabel is TRUE and true_eta is provided, sets
automated active updating scheme. Can be "ratio", in which case it
decides based on the ratio of the true etas, or "ranked", in which
case it decides based on the raw values of the true eta.}

\item{verbose}{Boolean deciding whether to print for debug.}

\item{dfm}{Quanteda document feature matrix.}

\item{keywords_list}{Existing list of keywords.}
}
\value{
A vector of words.
}
\description{
Actively query eta parameter.
}
