% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{query_label}
\alias{query_label}
\title{Label Query}
\usage{
query_label(
  docs,
  label_id_vec,
  n_class,
  labels,
  doc_name,
  index_name,
  labels_name = NULL,
  active_iter = NULL,
  maxIter = NULL,
  handlabel = TRUE,
  metadata_vars = NA,
  already_selected = NA,
  param_to_save = NA
)
}
\arguments{
\item{docs}{[matrix]    Matrix of labeled and/or unlabeled documents.}

\item{label_id_vec}{[vector]    Matrix of documents to be labeled.}

\item{n_class}{[numeric]   Number of classes to be considered.}

\item{labels}{[vector]    Vector of character strings indicating classification options.}

\item{doc_name}{[character] Character string indicating the variable in \code{docs} that
denotes the text of the documents to be classified.}

\item{index_name}{[character] Character string indicating the variable in 'docs' that
denotes the index value of the document to be classified.}

\item{labels_name}{[character] Character string indicating the variable in \code{docs}
that denotes the already known labels of the documents.
By default, value is set to \code{NULL}.}

\item{active_iter}{[numeric]   Numeric value denoting which iteration of the active learning cycle
the algorithm is in. Appears as header information to the user-labeling
process.}

\item{maxIter}{[numeric]   Numeric value denoting the maximum number of active learning iterations.}

\item{handlabel}{[logical]   Boolean logical value indicating whether to initiate user-input script.
If set to \code{FALSE}, and if \code{labels_name} is provided, the script
queries the document label directly from the column denoted by \code{labels_name}.}

\item{param_to_save}{[list]    List of variables to save if user chooses to.}
}
\value{
[matrix] or [list]     If finishes to completion, structured matrix of labeled and unlabeled documents, updated with
                              labels for the documents in `label_id_vec`. Otherwise, a list of already hand labeled values
                              to save.
}
\description{
Queries documents for classification by oracle.
}
