% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{region_sample_edge}
\alias{region_sample_edge}
\title{Log-Ratio Region Sampling}
\usage{
region_sample_edge(docs, max_query, edge = T, regions = "both")
}
\arguments{
\item{docs}{[dataframe] Documents with log ratio and cumulative sum columns.}

\item{max_query}{[integer] Number of documents to be queried.}

\item{edge}{[logical] Whether we sapmle from the edge of the posterior distributions as well}
}
\value{
Sampled documents.
}
\description{
Log-Ratio Region Sampling
}
