% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{reshapedocs}
\alias{reshapedocs}
\title{reshape docs object such that the prediction at each iteration is stored}
\usage{
reshapedocs(output)
}
\arguments{
\item{"output"}{is the entire output (not "output" element of the list) from the run_models_fast function}
}
\description{
reshape docs object such that the prediction at each iteration is stored
}
