% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{reshapedocs_helper}
\alias{reshapedocs_helper}
\title{(helper) reshape docs object such that the prediction at each iteration is stored}
\usage{
reshapedocs_helper(output_model)
}
\description{
(helper) reshape docs object such that the prediction at each iteration is stored
}
