% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{run_models_fast}
\alias{run_models_fast}
\title{Run Models Fast}
\usage{
run_models_fast(
  docs,
  docName,
  index_name,
  labelsName,
  n_class = 2,
  iters = 1,
  n_cluster = 2,
  trainSize,
  labelSize,
  querySize = 100,
  initSize = 100,
  lambda = 0.1,
  seed = NULL,
  useLabeled = F,
  queryType = "basic_entropy",
  fixed_words = NULL,
  dfm = NULL,
  verbose = T,
  forceList = T,
  bound = 0,
  weight_fpfn = F,
  log_ratio_conv_type = "none",
  mu = 1e-04,
  tau = 1e-04,
  regions = "both",
  lambda_decay = F,
  ld_rate = 0.2,
  init_vec = NULL,
  ...
)
}
\arguments{
\item{docs}{Matrix of documents.}

\item{docName}{String indicating column name containing document text.}

\item{index_name}{String indicating column name containing document index.}

\item{labelsName}{String indicating column name containing document labels.}

\item{n_class}{Integer indicating number of classes to use.}

\item{iters}{Number of monte-carlo iterations, the results of which are averaged over.}

\item{n_cluster}{Int; number of clusters to use.}

\item{trainSize}{Integer indicating number of documents to set aside as training data.}

\item{labelSize}{Integer indicating total number of documents to actively label.}

\item{querySize}{Integer indicating the number of documents to be labeled on each active iteration.}

\item{initSize}{Integer indicating how many documents to initialize the experiments with when `init_vec == NULL'.}

\item{lambda}{Numeric value indicating lambda value to use for "wEM" and "awEM" models.}

\item{seed}{Optional integer value to specify seed.}

\item{useLabeled}{Boolean used to indicate whether to calculate in-sample statistics using labeled docs or not. F by default.}

\item{queryType}{String indicating which type of uncertainty sampling scheme to use. "basic_entropy" by default.
Other options include "random" for random sampling, and "log_ratio" for log ratio sampling.}

\item{fixed_words}{List of words which remain fixed at their empirical probabilities across all active iteration. Experimental feature.}

\item{dfm}{Optional argument to supply an externally generaged dfm.}

\item{verbose}{Boolean used to indicate whether to show progress over iterations.}

\item{forceList}{Boolean used to indicate whether to sample randomly if insufficient entropy to use entropy sampling.}

\item{bound}{Float; if using entropy sampling, sets the lower bound of entropy which are chosen for sampling. e.g., if sampling
documents with bound 0.1, the algorithm will not sample documents with an entropy value below 0.1.}

\item{weight_fpfn}{Boolean; used to indicate whether to weight results by false positive and false negative
values if using log ratio sampling.}

\item{log_ratio_conv_type}{String; determines whether log ratio sampling if enabled has a convergence condition.
By default set to "none".}

\item{mu}{Float; parameter for log ratio sampling.}

\item{tau}{Float; parameter for log ratio sampling.}

\item{regions}{String; determines which regions in log ratio sampling scheme are accessed. By default set to "both".}

\item{lambda_decay}{Determines whether lambda value decays over active learning iterations or not.}

\item{ld_rate}{If `lambda_decay == TRUE`, sets the rate at which decay occurs.}

\item{...}{Optional arguments to pass to `get_dfm()` if a dfm object is not already supplied.}

\item{init_index}{Vector of index values to initialize experiments with.}
}
\value{
List of matrices containing model validation scores across each active iteration.
Outputs accuracy, precision, recall, and F1 scores for in-sample and out-of-sample validation.
Also outputs metadata from experiment that can be fed into the modeling functions `get_figures_grid()`
for seamless model labels.
---------------------------- WARNINGS -------------------------------- #
---------------------------- END WARNINGS ------------------------------ #
----------------------------  SETUP  ----------------------------------- #
create arrays if necessary
--------------------------  END SETUP  --------------------------------- #
--------------------------  MAIN LOOP  --------------------------------- #
set seed
seperate training from testing
Create containers for results
Use FBM for parallelization
Create DFM if necessary
Get initial documents index
Initial documents
parallelized dopar loop
Specify weight
specify querytype
specify lambda_decay val for weighted models
Run model
Assign results
weighted FP and FN to compute precision and recall
Get in-sample validation statistics
Get out-of-sample validation statistics
Get in-sample validation statistics
Get out-of-sample validation statistics
--------------------------  END MAIN LOOP  --------------------------------- #
--------------------------  RESHAPE DATA  ---------------------------------- #
If running monte carlo, impute results to array
clean output
Remove last column for in-label validation if hand labeled docs are not used for validation
--------------------------  END RESHAPE DATA  ------------------------------------------------- #
}
\description{
Runs 6 model specifications in parallel in order to quickly compare results across models.
}
