% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{run_models_fast_multi}
\alias{run_models_fast_multi}
\title{Run Models Fast}
\usage{
run_models_fast_multi(
  docs,
  docName,
  index_name,
  labelsName,
  n_class = 2,
  n_cluster = 2,
  iters = 1,
  trainSize,
  labelSize,
  querySize = 100,
  lambda = 0.1,
  seed = NULL,
  useLabeled = F,
  queryType = "basic_entropy",
  fixed_words = NULL,
  dfm = NULL,
  verbose = T,
  forceList = T,
  bound = 0,
  weight_fpfn = F,
  log_ratio_conv_type = "none",
  mu = 1e-04,
  tau = 1e-04,
  ...
)
}
\arguments{
\item{docs}{Matrix of documents.}

\item{docName}{String indicating column name containing document text.}

\item{index_name}{String indicating column name containing document index.}

\item{labelsName}{String indicating column name containing document labels.}

\item{n_class}{Integer indicating number of classes to use.}

\item{n_cluster}{Integer indicating number of clusters to use.}

\item{trainSize}{Integer indicating number of documents to set aside as training data.}

\item{labelSize}{Integer indicating total number of documents to actively label.}

\item{querySize}{Integer indicating the number of documents to be labeled on each active iteration.}

\item{lambda}{Numeric value indicating lambda value to use for "wEM" and "awEM" models.}

\item{seed}{Optional integer value to specify seed.}

\item{useLabeled}{Boolean used to indicate whether to calculate in-sample statistics using labeled docs or not. F by default.}

\item{queryType}{String indicating which type of uncertainty sampling scheme to use. "basic_entropy" by default.}

\item{fixed_words}{List of words which remain fixed at their empirical probabilities across all active iteration. Experimental feature.}

\item{dfm}{Optional argument to supply an externally generaged dfm.}

\item{verbose}{Boolean used to indicate whether to show progress over iterations.}

\item{...}{Optional arguments to pass to `get_dfm()` if a dfm object is not already supplied.}
}
\value{
List of matrices containing model validation scores across each active iteration.
Outputs accuracy, precision, recall, and F1 scores for in-sample and out-of-sample validation.
}
\description{
Runs 6 model specifications in parallel in order to quickly compare results across models.
}
