% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{run_unsupervised_EM}
\alias{run_unsupervised_EM}
\title{Run Unsupervised EM}
\usage{
run_unsupervised_EM(
  docs,
  doc_name,
  docsType,
  index_name,
  labelsName,
  dfm = NULL,
  extreme_words_pos = NULL,
  extreme_words_neg = NULL,
  stem = T,
  trimPct = 0,
  min_doc_freq = 1,
  idfWeight = F,
  ngrams = 1,
  n_extreme = 50,
  n_class = 2,
  export_plot = T,
  export_all = T,
  class_prob = NULL,
  word_prob = NULL,
  ...
)
}
\value{
Document level probabilities, word probabilities, scatterplot of word probablities.
}
\description{
Runs supervised EM on a corpus of documents cleaned by \code{clean_data()}.
}
