% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{sample_data}
\alias{sample_data}
\title{get a sample of dataset with a specific proportion of positive documents.}
\usage{
sample_data(data, n = NULL, pos_prob)
}
\description{
get a sample of dataset with a specific proportion of positive documents.
}
