% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{split_dfm}
\alias{split_dfm}
\title{Splits Document Term Matrix into two parts.}
\usage{
split_dfm(dfm, splitIndex)
}
\arguments{
\item{dfm}{[matrix]    Output of \code{get_dfm}. Document-Term Matrix.}

\item{splitIndex}{[vector]    Vector of strings denoting the row values to split on.}
}
\value{
[list]      List of two parts of the document term matrix.
}
\description{
Splits a document term matrix according to the provided \code{splitIndex}. Outputs a list of
             two sections: the first that belongs to the \code{splitIndex}, and the second that does not.
}
