% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EM_helper.R
\name{split_sparsemat}
\alias{split_sparsemat}
\title{Split DTM into positive and negative label areas
Extract sparse matrix triplet (row, col, value)}
\usage{
split_sparsemat(.C_train, .D_train)
}
\description{
Split DTM into positive and negative label areas
Extract sparse matrix triplet (row, col, value)
}
