% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tuning.R
\name{tune_lambda}
\alias{tune_lambda}
\title{Tune Lambda Weight and Number of Clusters}
\usage{
tune_lambda(
  docs,
  docName = "text",
  index_name = "id",
  labelsName = "label",
  n_clusters = c(2),
  maxQuery = 1,
  active = FALSE,
  lambdas,
  seed,
  prop_init = 0.1,
  parallel = TRUE,
  k = 5
)
}
\arguments{
\item{docs}{Dataframe for courpus.}

\item{docName}{String; name of variable for document text.}

\item{index_name}{String; name of variable for corpus index.}

\item{labelsName}{String; name of variable for document labels.}

\item{n_clusters}{vec; sequence of number of clusters to be tested.}

\item{maxQuery}{int; max number of documents ot query per iteration.}

\item{lambdas}{vec; sequence of lambda values to test.}

\item{seed}{int; Seed value for randomization.}

\item{parallel}{logical; Whether or not to parallelize the tuning process.}

\item{k}{int; number for cross-validation.}

\item{prop_init;}{float; Proportion of docs in each training set
to directly label; the rest are used by the EM.}
}
\value{
Results from experiment.
parallel
select best # of clusters and lambda in terms of the mean F1
}
\description{
Experiment function for tuning lambda.
}
