% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tuning.R
\name{tune_lambda_helper}
\alias{tune_lambda_helper}
\title{Helper Function for `Tune_Lambda()`}
\usage{
tune_lambda_helper(
  i,
  docs,
  docName,
  index_name,
  n_cluster,
  labelsName,
  initSize,
  initIndex,
  lambda,
  maxQuery,
  whichOutTest,
  maxActive,
  seed,
  verbose = TRUE
)
}
\arguments{
\item{i}{Iteration number.}

\item{docs}{Dataframe for courpus.}

\item{docName}{String; name of variable for document text.}

\item{index_name}{String; name of variable for corpus index.}

\item{n_cluster}{Int; number of clusters.}

\item{labelsName}{String; name of variable for document labels.}

\item{initSize}{Int; number of docs to initialize model with.}

\item{initIndex}{vec; Value to initialize model with.}

\item{lambda}{float; lambda value for weighting.}

\item{maxQuery}{int; max number of documents ot query per iteration.}

\item{whichOutTest}{vec; Index values to use for out-of-sample
validation.}

\item{seed}{int; Seed value for randomization.}

\item{verbose}{logical; Denotes whether model progress is printed.}
}
\value{
Statistics summary for active_EM() iteration.
}
\description{
helper function for tune_lambda()
}
