% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{tune_lambda_in_active}
\alias{tune_lambda_in_active}
\title{Tunes lambda value at each iteraction, if enabled}
\usage{
tune_lambda_in_active(
  docs,
  index_name,
  hand_labeled_index,
  n_cluster,
  tune_lambda_range,
  tune_lambda_prop_init,
  tune_lambda_parallel,
  tune_lambda_k,
  seed
)
}
\description{
Tunes lambda value at each iteraction, if enabled
}
