% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{tuning_algorithm}
\alias{tuning_algorithm}
\title{Active EM Parameter Tuning Algorithm}
\usage{
tuning_algorithm(
  docs,
  docName,
  index_name,
  labelsName,
  measure = "accuracy",
  max_runs = 30,
  seed = 123,
  stem = c(F, T),
  ngrams = c(1:3),
  trimPct = c(0, 1e-04, 0.001, 0.01),
  minDocFreq = c(1, 2, 3),
  idfWeight = c(F, T),
  removeStopWords = c(F, T),
  minChar = 1:4,
  weight = seq(0, 1, 0.1),
  bound = seq(0, 1, 0.1),
  initSize = 5,
  maxActive = 20,
  maxQuery = 1,
  queryType = c("basic_entropy", "random"),
  forceSweep = NA,
  dfmOut = F
)
}
\description{
Tries to find optimal combination of pre-processing and active algorithm parameters.
}
