% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_active_helper.R
\name{update_results}
\alias{update_results}
\title{Update model results.}
\usage{
update_results(
  include_out_stats,
  res_obj,
  agg_output_in,
  agg_output_out,
  hl_index,
  i,
  time_sec
)
}
\description{
updates model results
}
