% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{word_scatterplots}
\alias{word_scatterplots}
\title{Get scatterplots showing the distribution of word parameters.}
\usage{
word_scatterplots(
  docs,
  docsType,
  extreme_words_pos = NULL,
  extreme_words_neg = NULL,
  fixed_words = NULL,
  n_extreme = 50,
  ...
)
}
\description{
Run models according to specifications and generate scatterplots
from word probabilities.
}
